Added structure of form
There is HTML5 code. I created a form in order to typing informations of books which user want to borrow in a library. Form contains 4 inputs. One for title of book, one for author of book, another one for making a prioryty of reading and the last one for choosing a category of book.  User can submit his inputed information by pushing button which is placed below inputs.